% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_vioplot.R
\name{gender_vioplot}
\alias{gender_vioplot}
\title{The sina weibo dot chart}
\usage{
gender_vioplot(df_comments, title_text = "Gender VS Rank")
}
\arguments{
\item{df_comments}{the dataframe of weibo comments}

\item{stopwords}{the stop words}
}
\description{
The sina weibo dot chart
}
\examples{
df_comments = get_comments(access_token, weiboid)
gender_vioplot(df_comments)
}

