% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Get comments of weibo}
\usage{
get_comments(access_token, weiboid, maxpage = 5)
}
\arguments{
\item{access_token}{get from http://open.weibo.com/tools/console}

\item{weiboid}{the id of weibo}

\item{maxpage}{the max page of weibo comments}
}
\description{
This function get comments of weibo
}

