% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get fields of weibo comments}
\usage{
get_fields(lst)
}
\arguments{
\item{access_token}{Acess Token}

\item{weiboid}{the id of weibo}
}
\description{
This function get fields of weibo comments
}
\examples{
get_fields(comments)
}

