% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province_dotchart.R
\name{province_dotchart}
\alias{province_dotchart}
\title{The sina weibo dot chart}
\usage{
province_dotchart(df_comments, stopprovince = c(""), maxrow = 100,
  title_text = "Users province distribution")
}
\arguments{
\item{df_comments}{the dataframe of weibo comments}

\item{stopprovince}{the stop provinces}
}
\description{
The sina weibo dot chart
}
\examples{
df_comments = get_comments(access_token, weiboid, maxpage = 10)
province_dotchart(df_comments)
}

