% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_vioplot.R
\name{verify_vioplot}
\alias{verify_vioplot}
\title{The sina weibo dot chart}
\usage{
verify_vioplot(df_comments, title_text = "Verify VS Rank")
}
\arguments{
\item{df_comments}{the dataframe of weibo comments}

\item{stopwords}{the stop words}
}
\description{
The sina weibo dot chart
}
\examples{
df_comments = get_comments(access_token, weiboid)
verify_vioplot(df_comments)
}

