% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibo_cloud.R
\name{weibo_cloud}
\alias{weibo_cloud}
\title{The sina weibo wordcloud}
\usage{
weibo_cloud(df_comments, stopwords = c(""), maxrow = 100)
}
\arguments{
\item{df_comments}{the dataframe of weibo comments}

\item{stopwords}{the stop words}
}
\description{
This function generates wordcloud of weibo comments.
}
\examples{
df_comments = get_comments(access_token, weiboid, maxpage = 10)
weibo_cloud(df_comments)
}

