% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weiboda.R
\name{weiboda}
\alias{weiboda}
\title{The sina weibo data analysis}
\usage{
weiboda(df_comments)
}
\arguments{
\item{df_comments}{the dataframe of weibo comments}
}
\description{
Analysis the weibo comments.
}
\examples{
library(rwda)
df_comments = get_comments(access_token, weiboid)
weiboda(df_comments)
}

